refactor: Check lockfile in update.ts
