+++ src/main/update/launcher-update.ts
@@ -1,3 +1,4 @@
+import path from "path";
 import { DownloadItem, app } from "electron";
 import { download, Options as ElectronDLOptions } from "electron-dl";
 import extractZip from "extract-zip";
@@ -6,20 +7,10 @@ import { IDownloadProgress } from "src/interfaces/ipc";
 import { tmpName } from "tmp-promise";
 import { DownloadBinaryFailedError } from "../exceptions/download-binary-failed";
 import { get as getConfig } from "../../config";
-import path from "path";
 import fs from "fs";
-import lockfile from "lockfile";
 import { spawn as spawnPromise } from "child-process-promise";
-import { playerUpdate } from "./player-update";
 import { IUpdate } from "./check";
-
-const lockfilePath = path.join(path.dirname(app.getPath("exe")), "lockfile");
-
-export interface IUpdateOptions {
-  downloadStarted(): Promise<void>;
-  relaunchRequired(): void;
-  getWindow(): Electron.BrowserWindow | null;
-}
+import { IUpdateOptions } from "./update";
 
 export async function launcherUpdate(
   update: IUpdate,
@@ -35,24 +26,6 @@ export async function launcherUpdate(
   if (win === null) {
     console.log("Stop update process because win is null.");
     return;
-  }
-
-  if (lockfile.checkSync(lockfilePath)) {
-    console.log(
-      "'encounter different version' event seems running already. Stop this flow."
-    );
-    return;
-  }
-
-  try {
-    lockfile.lockSync(lockfilePath);
-    console.log(
-      "Created 'encounter different version' lockfile at ",
-      lockfilePath
-    );
-  } catch (e) {
-    console.error("Error occurred during trying lock.");
-    throw e;
   }
 
   await listeners.downloadStarted();
@@ -191,72 +164,4 @@ export async function launcherUpdate(
     "The existing and new configuration files has been merged:",
     config
   );
-
-  lockfile.unlockSync(lockfilePath);
-  console.log(
-    "Removed 'encounter different version' lockfile at ",
-    lockfilePath
-  );
-  /*
-      autoUpdater provided from Electron must code signed to work at macOS
-      So we can't use it for now.
-
-      FIXME: By attatching Squirell, make it update, later.
-
-      const { path: tmpPath } = await tmp.file({
-        postfix: ".json",
-        discardDescriptor: true,
-      });
-      const tmpFile = await fs.promises.open(tmpPath, "w");
-      const feedData = {
-        url: downloadUrl,
-      };
-
-      await tmpFile.writeFile(JSON.stringify(feedData), "utf8");
-      console.log(`Wrote a temp feed JSON file:`, tmpPath);
-      autoUpdater.setFeedURL({ url: `file://${tmpPath}` });
-
-      autoUpdater.on("error", (message) =>
-        console.error("AUTOUPDATER:ERROR", message)
-      );
-      autoUpdater.on("checking-for-update", () =>
-        console.error("AUTOUPDATER:CHECKING-FOR-UPDATE")
-      );
-      autoUpdater.on("update-available", () =>
-        console.error("AUTOUPDATER:UPDATE-AVAILABLE")
-      );
-      autoUpdater.on("update-not-available", () =>
-        console.error("AUTOUPDATER:UPDATE-NOT-AVAILABLE")
-      );
-      autoUpdater.on(
-        "update-downloaded",
-        (event, releaseNotes, releaseName, releaseDate, updateURL) =>
-          console.error(
-            "AUTOUPDATER:UPDATE-DOWNLOADED",
-            event,
-            releaseNotes,
-            releaseName,
-            releaseDate,
-            updateURL
-          )
-      );
-      autoUpdater.on("before-quit-for-update", () =>
-        console.error("AUTOUPDATER:BEFORE-QUIT-FOR-UPDATE")
-      );
-
-      autoUpdater.checkForUpdates();
-      */
-}
-
-export function isUpdating() {
-  return lockfile.checkSync(lockfilePath);
-}
-
-/**
- * unlock if lockfile locked.
- */
-export function cleanUpLockfile() {
-  if (lockfile.checkSync(lockfilePath)) {
-    lockfile.unlockSync(lockfilePath);
-  }
 }
