+++ src/main/update/player-update.ts
@@ -2,18 +2,11 @@ import { DownloadItem, app } from "electron";
 import { download, Options as ElectronDLOptions } from "electron-dl";
 import { IDownloadProgress } from "src/interfaces/ipc";
 import { DownloadBinaryFailedError } from "../exceptions/download-binary-failed";
-import path from "path";
 import fs from "fs";
 import extractZip from "extract-zip";
 import { spawn as spawnPromise } from "child-process-promise";
 import { cleanupOldPlayer } from "./util";
 import { playerPath } from "../../config";
-import lockfile from "lockfile";
-
-const lockfilePath = path.join(
-  path.dirname(app.getPath("exe")),
-  "player-lockfile"
-);
 
 export async function playerUpdate(
   downloadUrl: string,
@@ -22,24 +15,6 @@ export async function playerUpdate(
   if (win === null) {
     console.log("Stop update process because win is null.");
     return;
-  }
-
-  if (lockfile.checkSync(lockfilePath)) {
-    console.log(
-      "[player] 'encounter different version' event seems running already. Stop this flow."
-    );
-    return;
-  }
-
-  try {
-    lockfile.lockSync(lockfilePath);
-    console.log(
-      "[player] Created 'encounter different version' lockfile at ",
-      lockfilePath
-    );
-  } catch (e) {
-    console.error("[player] Error occurred during trying lock.");
-    throw e;
   }
 
   win.webContents.send("update player download started");
@@ -131,10 +106,4 @@ export async function playerUpdate(
   win.webContents.send("update player extract complete");
 
   await fs.promises.unlink(dlPath);
-
-  lockfile.unlockSync(lockfilePath);
-  console.log(
-    "[player] Removed 'encounter different version' lockfile at ",
-    lockfilePath
-  );
 }
